Repackage migration interfaces
