+++ api/src/org/labkey/api/files/FileSystemAttachmentType.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2017 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.labkey.api.files;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.attachments.AttachmentType;
+import org.labkey.api.data.CoreSchema;
+import org.labkey.api.data.SQLFragment;
+
+public class FileSystemAttachmentType implements AttachmentType
+{
+    private static final FileSystemAttachmentType INSTANCE = new FileSystemAttachmentType();
+
+    public static FileSystemAttachmentType get()
+    {
+        return INSTANCE;
+    }
+
+    private FileSystemAttachmentType()
+    {
+    }
+
+    @Override
+    public @NotNull String getUniqueName()
+    {
+        return getClass().getName();
+    }
+
+    @Override
+    public @Nullable SQLFragment getSelectParentEntityIdsSql()
+    {
+        return new SQLFragment("SELECT EntityId FROM ").append(CoreSchema.getInstance().getMappedDirectories(), "md");
+    }
+}
