+++ assay/src/org/labkey/assay/AssayResultMigrationSchemaHandler.java
@@ -1,10 +1,8 @@
 package org.labkey.assay;
 
+import org.apache.commons.lang3.Strings;
 import org.apache.logging.log4j.Logger;
 import org.labkey.api.assay.AbstractTsvAssayProvider;
-import org.labkey.api.data.DatabaseMigrationService.DataFilter;
-import org.labkey.api.data.DatabaseMigrationService.DefaultMigrationSchemaHandler;
-import org.labkey.api.data.DatabaseMigrationService.ExperimentDeleteService;
 import org.labkey.api.data.DbSchema;
 import org.labkey.api.data.DbSchemaType;
 import org.labkey.api.data.SQLFragment;
@@ -14,8 +12,12 @@ import org.labkey.api.data.SimpleFilter.OrClause;
 import org.labkey.api.data.SimpleFilter.SQLClause;
 import org.labkey.api.data.SqlSelector;
 import org.labkey.api.data.TableInfo;
+import org.labkey.api.migration.DatabaseMigrationService.DataFilter;
+import org.labkey.api.migration.DefaultMigrationSchemaHandler;
+import org.labkey.api.migration.ExperimentDeleteService;
 import org.labkey.api.util.GUID;
 import org.labkey.api.util.logging.LogHelper;
+import org.labkey.assay.plate.PlateReplicateStatsDomainKind;
 
 import java.util.Collection;
 import java.util.Set;
@@ -29,18 +31,18 @@ class AssayResultMigrationSchemaHandler extends DefaultMigrationSchemaHandler
         super(DbSchema.get(AbstractTsvAssayProvider.ASSAY_SCHEMA_NAME, DbSchemaType.Provisioned));
     }
 
-    // Provisioned assay result tables occasionally have no DataId column; hopefully they have an LSID column.
-    private boolean hasDataIdColumn(TableInfo sourceTable)
+    private boolean skipTable(TableInfo sourceTable)
     {
-        return sourceTable.getColumn("DataId") != null;
+        // For now, we're ignoring this table since it's empty in our first migration client's database
+        return Strings.CI.endsWith(sourceTable.getName(), PlateReplicateStatsDomainKind.ASSAY_PLATE_REPLICATE);
     }
 
     @Override
     public FilterClause getContainerClause(TableInfo sourceTable, Set<GUID> containers)
     {
-        return new SQLClause(
-            new SQLFragment(hasDataIdColumn(sourceTable) ? "DataId IN (SELECT RowId" : "LSID IN (SELECT LSID")
-                .append(" FROM exp.Data WHERE Container")
+        return new SQLClause(skipTable(sourceTable) ?
+            new SQLFragment("1 = 0") :
+            new SQLFragment("DataId IN (SELECT RowId FROM exp.Data WHERE Container")
                 .appendInClause(containers, sourceTable.getSqlDialect())
                 .append(")")
         );
@@ -55,26 +57,27 @@ class AssayResultMigrationSchemaHandler extends DefaultMigrationSchemaHandler
     @Override
     public void afterTable(TableInfo sourceTable, TableInfo targetTable, SimpleFilter notCopiedFilter)
     {
-        SQLFragment objectIdSql = new SQLFragment("SELECT ObjectId FROM exp.Data WHERE ")
-            .append(hasDataIdColumn(sourceTable) ? "RowId IN (SELECT DataId" : "LSID IN (SELECT LSID")
-            .append(" FROM ")
-            .appendIdentifier(sourceTable.getSelectName())
-            .append(" ")
-            .append(notCopiedFilter.getSQLFragment(sourceTable.getSqlDialect()))
-            .append(")");
-
-        Collection<Long> notCopiedObjectIds = new SqlSelector(sourceTable.getSchema(), objectIdSql).getCollection(Long.class);
-
-        if (notCopiedObjectIds.isEmpty())
+        if (!skipTable(sourceTable))
         {
-            LOG.info(rowsNotCopied(0));
-        }
-        else
-        {
-            LOG.info("{} -- deleting associated rows from exp.Data, exp.Object, etc.", rowsNotCopied(notCopiedObjectIds.size()));
+            SQLFragment objectIdSql = new SQLFragment("SELECT ObjectId FROM exp.Data WHERE RowId IN (SELECT DataId FROM ")
+                .appendIdentifier(sourceTable.getSelectName())
+                .append(" ")
+                .append(notCopiedFilter.getSQLFragment(sourceTable.getSqlDialect()))
+                .append(")");
 
-            // Delete exp.Data, exp.Object, etc. rows associated with the rows that weren't copied
-            ExperimentDeleteService.get().deleteDataRows(notCopiedObjectIds);
+            Collection<Long> notCopiedObjectIds = new SqlSelector(sourceTable.getSchema(), objectIdSql).getCollection(Long.class);
+
+            if (notCopiedObjectIds.isEmpty())
+            {
+                LOG.info(rowsNotCopied(0));
+            }
+            else
+            {
+                LOG.info("{} -- deleting associated rows from exp.Data, exp.Object, etc.", rowsNotCopied(notCopiedObjectIds.size()));
+
+                // Delete exp.Data, exp.Object, etc. rows associated with the rows that weren't copied
+                ExperimentDeleteService.get().deleteDataRows(notCopiedObjectIds);
+            }
         }
     }
 }
