+++ experiment/src/org/labkey/experiment/ExperimentModule.java
@@ -30,7 +30,6 @@ import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
 import org.labkey.api.data.ContainerManager;
 import org.labkey.api.data.CoreSchema;
-import org.labkey.api.data.DatabaseMigrationService;
 import org.labkey.api.data.DbSchema;
 import org.labkey.api.data.JdbcType;
 import org.labkey.api.data.NameGenerator;
@@ -73,6 +72,9 @@ import org.labkey.api.exp.xar.LSIDRelativizer;
 import org.labkey.api.exp.xar.LsidUtils;
 import org.labkey.api.files.FileContentService;
 import org.labkey.api.files.TableUpdaterFileListener;
+import org.labkey.api.migration.AssaySkipFilter;
+import org.labkey.api.migration.DatabaseMigrationService;
+import org.labkey.api.migration.ExperimentDeleteService;
 import org.labkey.api.module.ModuleContext;
 import org.labkey.api.module.ModuleLoader;
 import org.labkey.api.module.SpringModule;
@@ -877,7 +879,8 @@ public class ExperimentModule extends SpringModule
         DatabaseMigrationService.get().registerSchemaHandler(new SampleTypeMigrationSchemaHandler());
         DataClassMigrationSchemaHandler dcHandler = new DataClassMigrationSchemaHandler();
         DatabaseMigrationService.get().registerSchemaHandler(dcHandler);
-        DatabaseMigrationService.ExperimentDeleteService.setInstance(dcHandler);
+        ExperimentDeleteService.setInstance(dcHandler);
+        DatabaseMigrationService.get().registerMigrationFilter(new AssaySkipFilter());
     }
 
     @Override
